Make dontExpectToFail to check if test should've failed before call
dontExpectToFail should check if the test should fail before calling the
method. Otherwise the tests may let tests pass even when the method
expected to fail doesn't fail. This happened because the mode was changed
from expecting errors to not expecting errors at the end of the test. So
before switching mode it has to be checked if the test should've failed.