Fix bug with ExpectCall.notToFail expecation
When notToFail was used, the users that were not defined, always were
considered also not to fail. They should, ofcourse, consireded to fail.
Removed possibility to mix fail and not fail style methods since it would
make no sense to allow user to make toFail(...).notToFail(...) calls. The
required cases should be possible to write with fail or not fail style
calls.