Initial work for component based tests
Started to work on a major refactor and archtectural change which makes
possible to use MUTR library via a class property rather than inheritance
and customized JUnit runner. This change will eventually enable better
Spring integration since it doesn't require a custom Spring support.
The Spring support can be implemented by the library user by using
JUnit rules.

Removed examples for now. At the time of rebasing on top of origin/master
the examples for 0.5.0 are in a separate project. Those tests work on later
commits but not the one here. Therefore the examples are only introduced in
a later commit where the tests are more meaningful.